+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Fix windows build.
 -   Return an error when creating a halight swap if lnd certificate or macaroon are unavailable instead of failing silently.
 
+## Changed
+
+-   **Breaking Change**: Replace `Parity` config section with `web3_url`.
+
 ## [0.7.3] - 2020-04-14
 
 ### Fixed